resolve conflicts in restructuring the script. Fixes #22
