+++ setup.py
@@ -271,7 +271,10 @@ else: # OSX
 ext_modules = [
     Extension(
         'BanditPAM',
-        [os.path.join('src', 'kmedoids_pywrapper.cpp'), os.path.join('src', 'kmedoids_ucb.cpp')],
+        [os.path.join('src', 'kmedoids_pywrapper.cpp'), 
+         os.path.join('src', 'kmedoids_algorithm.cpp'),
+         os.path.join('src', 'pam.cpp'),
+         os.path.join('src', 'banditpam.cpp')],
         include_dirs=include_dirs,
         library_dirs=[
             '/usr/local/lib',
@@ -308,5 +311,5 @@ setup(
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
-    headers=[os.path.join('headers', 'kmedoids_ucb.hpp')],
+    headers=[os.path.join('headers', 'kmedoids_algorithm.hpp')],
 )
