+++ src/main.cpp
@@ -9,7 +9,7 @@
  * ./src/build/BanditPAM -f [path/to/input] -k [number of clusters]
  */
 
-#include "kmedoids_ucb.hpp"
+#include "kmedoids_algorithm.hpp"
 
 #include <armadillo>
 #include <chrono>
